updated anomaly detection, more tweaks
