+++ docs/tutorials/anomaly-detection/tutorial_app.py
@@ -2,14 +2,21 @@ import os
 import random
 import time
 
-from quixstreams import Application
+from quixstreams.sources import Source
 
 
-class TemperatureEventGenerator:
+class TemperatureEventGenerator(Source):
     """
-    Generates temperature readings for three different machines.
+    What a Source is:
+    A Quix Streams Source enables Applications to read data from something other than a Kafka topic.
+    Basically, this generates JSON data that the StreamingDataFrame consumes directly.
+    This simplifies the example by not having to run both a producer and Application.
 
-    Machine ID's 0, 1 are functioning normally, 2 is malfunctioning.
+    What it does:
+    Generates temperature readings for three different machines.
+    Machine ID's 0 and 1 are functioning normally; ID 2 is malfunctioning.
+    It cycles through the ID's, generating random temperature changes of -1, 0, or 1.
+    These probabilities change depending on current temperature and operation.
     """
 
     # Probabilities are % chance for [-1, 0, 1] change given a current temp in the
@@ -33,7 +40,6 @@ class TemperatureEventGenerator:
     }
 
     def __init__(self):
-        self.stop = False
         self.event_count = 0
         self.machine_temps = {0: 66, 1: 58, 2: 62}
         self.machine_types = {
@@ -41,6 +47,7 @@ class TemperatureEventGenerator:
             1: self.probabilities_normal,
             2: self.probabilities_issue,
         }
+        super().__init__(name="temperature-event-generator")
 
     def update_machine_temp(self, machine_id):
         """
@@ -57,9 +64,6 @@ class TemperatureEventGenerator:
 
         It tracks/loops through the Machine ID's for equal number of readings.
         """
-        if self.stop:
-            return None
-
         machine_id = self.event_count % 3
         self.update_machine_temp(machine_id)
         event_out = {
@@ -71,19 +75,45 @@ class TemperatureEventGenerator:
         }
         self.event_count += 1
         if self.machine_temps[machine_id] == 100:
-            self.stop = True
+            self.stop()
         return event_out
 
-
-_app = Application(broker_address=os.environ.get("BROKER_ADDRESS", "localhost:9092"))
-topic = _app.topic(name="temperature_readings")
-event_generator = TemperatureEventGenerator()
-
-if __name__ == "__main__":
-    with _app.get_producer() as producer:
-        while event := event_generator.generate_event():
-            event = topic.serialize(**event)
+    def run(self):
+        while self.running:
+            event = self.serialize(**self.generate_event())
             print(f"producing event for MID {event.key}, {event.value}")
-            producer.produce(key=event.key, value=event.value, topic=topic.name)
+            self.produce(key=event.key, value=event.value)
             time.sleep(0.2)  # just to make things easier to follow along
         print("An alert should have been generated by now; stopping producing.")
+
+
+def setup_and_run_application():
+    from quixstreams import Application
+
+    app = Application(
+        broker_address=os.environ.get("BROKER_ADDRESS", "localhost:9092"),
+        consumer_group="temperature_alerter",
+        auto_offset_reset="earliest",
+    )
+    alerts_topic = app.topic(name="alerts")
+
+    def should_alert(window_value: int, key, timestamp, headers):
+        if window_value >= 90:
+            print(f"Alerting for MID {key}: Average Temperature {window_value}")
+            return True
+
+    # If reading from a Kafka topic, pass topic=<Topic> instead of a source
+    sdf = app.dataframe(source=TemperatureEventGenerator())
+    sdf = sdf.apply(lambda data: data["Temperature_C"])
+    sdf = sdf.hopping_window(duration_ms=5000, step_ms=1000).mean().current()
+    sdf = sdf.apply(lambda result: round(result["value"], 2)).filter(
+        should_alert, metadata=True
+    )
+    # .to_topic() does not require reassignment ("in-place" operation), but does no harm
+    sdf = sdf.to_topic(alerts_topic)
+
+    app.run()
+
+
+if __name__ == "__main__":
+    setup_and_run_application()
